Stop using regexp2 / allow weaving under runtime / move under /runtime
