+++ _integration-tests/gls/access_test.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/DataDog/orchestrion/built"
+	"github.com/DataDog/orchestrion/runtime/built"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
