+++ go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/charmbracelet/lipgloss v1.0.0
 	github.com/dave/dst v0.27.3
 	github.com/dave/jennifer v1.7.1
-	github.com/dlclark/regexp2 v1.11.4
 	github.com/fsnotify/fsnotify v1.8.0
 	github.com/gohugoio/hugo v0.136.5
 	github.com/google/go-licenses/v2 v2.0.0-alpha.1
@@ -129,6 +128,7 @@ require (
 	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/disintegration/gift v1.2.1 // indirect
+	github.com/dlclark/regexp2 v1.11.4 // indirect
 	github.com/dustin/go-humanize v1.0.1 // indirect
 	github.com/eapache/go-resiliency v1.7.0 // indirect
 	github.com/eapache/go-xerial-snappy v0.0.0-20230731223053-c322873962e3 // indirect
