+++ runtime/doc.go
@@ -1,0 +1,10 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2023-present Datadog, Inc.
+
+// Package runtime provides the run-time API of orchestrion packages. It
+// includes features such as some run-time healpers used by injected
+// instrumentation code, as well as some facilities to perform telemetry or
+// ensure an application was indeed built with orchestrion.
+package runtime
